1. New API for ShareHolder
git-svn-id: http://192.168.110.32/svn/Products/PFF/git-migration/branches/mlod@124133 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/service/customermasters/impl/DirectorDetailServiceImpl.java
@@ -43,37 +43,50 @@
 
 package com.pennant.backend.service.customermasters.impl;
 
+import org.apache.commons.lang.StringUtils;
 import org.apache.log4j.Logger;
 import org.springframework.beans.BeanUtils;
 
+import com.pennant.app.util.ErrorUtil;
 import com.pennant.backend.dao.audit.AuditHeaderDAO;
+import com.pennant.backend.dao.customermasters.CustomerDocumentDAO;
 import com.pennant.backend.dao.customermasters.DirectorDetailDAO;
+import com.pennant.backend.dao.systemmasters.CityDAO;
+import com.pennant.backend.dao.systemmasters.GenderDAO;
+import com.pennant.backend.dao.systemmasters.NationalityCodeDAO;
+import com.pennant.backend.dao.systemmasters.SalutationDAO;
+import com.pennant.backend.model.audit.AuditDetail;
 import com.pennant.backend.model.audit.AuditHeader;
+import com.pennant.backend.model.customermasters.Customer;
 import com.pennant.backend.model.customermasters.DirectorDetail;
+import com.pennant.backend.model.systemmasters.City;
+import com.pennant.backend.model.systemmasters.NationalityCode;
 import com.pennant.backend.service.GenericService;
 import com.pennant.backend.service.customermasters.DirectorDetailService;
 import com.pennant.backend.service.customermasters.validation.CustomerDirectorValidation;
 import com.pennant.backend.util.PennantConstants;
+import com.pennanttech.pennapps.core.model.ErrorDetail;
 
 /**
  * Service implementation for methods that depends on <b>DirectorDetail</b>.<br>
  * 
  */
 public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> implements DirectorDetailService {
-
 	private static final Logger logger = Logger.getLogger(DirectorDetailServiceImpl.class);
 
 	private AuditHeaderDAO auditHeaderDAO;
 	private DirectorDetailDAO directorDetailDAO;
+	private CityDAO cityDAO;
+	private GenderDAO genderDAO;
+	private SalutationDAO salutationDAO;
+	private CustomerDocumentDAO customerDocumentDAO;
+	private NationalityCodeDAO nationalityCodeDAO;
+
 	private CustomerDirectorValidation customerDirectorValidation;
 
 	public DirectorDetailServiceImpl() {
 		super();
 	}
-
-	// ******************************************************//
-	// ****************** getter / setter *******************//
-	// ******************************************************//
 
 	public AuditHeaderDAO getAuditHeaderDAO() {
 		return auditHeaderDAO;
@@ -91,8 +104,39 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 		this.directorDetailDAO = directorDetailDAO;
 	}
 
-	public CustomerDirectorValidation getDirectorValidation() {
+	public CityDAO getCityDAO() {
+		return cityDAO;
+	}
 
+	public void setCityDAO(CityDAO cityDAO) {
+		this.cityDAO = cityDAO;
+	}
+
+	public NationalityCodeDAO getNationalityCodeDAO() {
+		return nationalityCodeDAO;
+	}
+
+	public void setNationalityCodeDAO(NationalityCodeDAO nationalityCodeDAO) {
+		this.nationalityCodeDAO = nationalityCodeDAO;
+	}
+
+	public GenderDAO getGenderDAO() {
+		return genderDAO;
+	}
+
+	public void setGenderDAO(GenderDAO genderDAO) {
+		this.genderDAO = genderDAO;
+	}
+
+	public void setSalutationDAO(SalutationDAO salutationDAO) {
+		this.salutationDAO = salutationDAO;
+	}
+
+	public void setCustomerDocumentDAO(CustomerDocumentDAO customerDocumentDAO) {
+		this.customerDocumentDAO = customerDocumentDAO;
+	}
+
+	public CustomerDirectorValidation getDirectorValidation() {
 		if (customerDirectorValidation == null) {
 			this.customerDirectorValidation = new CustomerDirectorValidation(directorDetailDAO);
 		}
@@ -110,12 +154,15 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 	}
 
 	/**
-	 * saveOrUpdate method method do the following steps. 1) Do the Business validation by using
-	 * businessValidation(auditHeader) method if there is any error or warning message then return the auditHeader. 2)
-	 * Do Add or Update the Record a) Add new Record for the new record in the DB table
-	 * CustomerDirectorDetail/CustomerDirectorDetail_Temp by using DirectorDetailDAO's save method b) Update the Record
-	 * in the table. based on the module workFlow Configuration. by using DirectorDetailDAO's update method 3) Audit the
-	 * record in to AuditHeader and AdtCustomerDirectorDetail by using auditHeaderDAO.addAudit(auditHeader)
+	 * saveOrUpdate method method do the following steps. 1) Do the Business
+	 * validation by using businessValidation(auditHeader) method if there is
+	 * any error or warning message then return the auditHeader. 2) Do Add or
+	 * Update the Record a) Add new Record for the new record in the DB table
+	 * CustomerDirectorDetail/CustomerDirectorDetail_Temp by using
+	 * DirectorDetailDAO's save method b) Update the Record in the table. based
+	 * on the module workFlow Configuration. by using DirectorDetailDAO's update
+	 * method 3) Audit the record in to AuditHeader and
+	 * AdtCustomerDirectorDetail by using auditHeaderDAO.addAudit(auditHeader)
 	 * 
 	 * @param AuditHeader
 	 *            (auditHeader)
@@ -151,10 +198,12 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 	}
 
 	/**
-	 * delete method do the following steps. 1) Do the Business validation by using businessValidation(auditHeader)
-	 * method if there is any error or warning message then return the auditHeader. 2) delete Record for the DB table
-	 * CustomerDirectorDetail by using DirectorDetailDAO's delete method with type as Blank 3) Audit the record in to
-	 * AuditHeader and AdtCustomerDirectorDetail by using auditHeaderDAO.addAudit(auditHeader)
+	 * delete method do the following steps. 1) Do the Business validation by
+	 * using businessValidation(auditHeader) method if there is any error or
+	 * warning message then return the auditHeader. 2) delete Record for the DB
+	 * table CustomerDirectorDetail by using DirectorDetailDAO's delete method
+	 * with type as Blank 3) Audit the record in to AuditHeader and
+	 * AdtCustomerDirectorDetail by using auditHeaderDAO.addAudit(auditHeader)
 	 * 
 	 * @param AuditHeader
 	 *            (auditHeader)
@@ -178,7 +227,8 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 	}
 
 	/**
-	 * getDirectorDetailById fetch the details by using DirectorDetailDAO's getDirectorDetailById method.
+	 * getDirectorDetailById fetch the details by using DirectorDetailDAO's
+	 * getDirectorDetailById method.
 	 * 
 	 * @param id
 	 *            (int)
@@ -192,8 +242,10 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 	}
 
 	/**
-	 * getApprovedDirectorDetailById fetch the details by using DirectorDetailDAO's getDirectorDetailById method . with
-	 * parameter id and type as blank. it fetches the approved records from the CustomerDirectorDetail.
+	 * getApprovedDirectorDetailById fetch the details by using
+	 * DirectorDetailDAO's getDirectorDetailById method . with parameter id and
+	 * type as blank. it fetches the approved records from the
+	 * CustomerDirectorDetail.
 	 * 
 	 * @param id
 	 *            (int)
@@ -204,15 +256,52 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 	}
 
 	/**
-	 * doApprove method do the following steps. 1) Do the Business validation by using businessValidation(auditHeader)
-	 * method if there is any error or warning message then return the auditHeader. 2) based on the Record type do
-	 * following actions a) DELETE Delete the record from the main table by using getDirectorDetailDAO().delete with
-	 * parameters directorDetail,"" b) NEW Add new record in to main table by using getDirectorDetailDAO().save with
-	 * parameters directorDetail,"" c) EDIT Update record in the main table by using getDirectorDetailDAO().update with
-	 * parameters directorDetail,"" 3) Delete the record from the workFlow table by using getDirectorDetailDAO().delete
-	 * with parameters directorDetail,"_Temp" 4) Audit the record in to AuditHeader and AdtCustomerDirectorDetail by
-	 * using auditHeaderDAO.addAudit(auditHeader) for Work flow 5) Audit the record in to AuditHeader and
-	 * AdtCustomerDirectorDetail by using auditHeaderDAO.addAudit(auditHeader) based on the transaction Type.
+	 * doApprove method do the following steps. 1) Do the Business validation by
+	 * using businessValidation(auditHeader) method if there is any error or
+	 * warning message then return the auditHeader. 2) based on the Record type
+	 * do following actions a) DELETE Delete the record from the main table by
+	 * using getDirectorDetailDAO().delete with parameters directorDetail,"" b)
+	 * NEW Add new record in to main table by using getDirectorDetailDAO().save
+	 * with parameters directorDetail,"" c) EDIT Update record in the main table
+	 * by using getDirectorDetailDAO().update with parameters directorDetail,""
+	 * 3) Delete the record from the workFlow table by using
+	 * getDirectorDetailDAO().delete with parameters directorDetail,"_Temp" 4)
+	 * Audit the record in to AuditHeader and AdtCustomerDirectorDetail by using
+	 * auditHeaderDAO.addAudit(auditHeader) for Work flow 5) Audit the record in
+	 * to AuditHeader and AdtCustomerDirectorDetail by using
+	 * auditHeaderDAO.addAudit(auditHeader) based on the transaction Type.
+	 * 
+	 */
+
+	@Override
+	public DirectorDetail getApprovedDirectorDetailByDirectorId(long directorId) {
+		return getDirectorDetailDAO().getDirectorDetailByDirectorId(directorId, "_AView");
+	}
+
+	/**
+	 * 
+	 */
+
+	@Override
+	public int getVersion(long custID, long directorId) {
+		return getDirectorDetailDAO().getVersion(custID, directorId);
+	}
+
+	/**
+	 * doApprove method do the following steps. 1) Do the Business validation by
+	 * using businessValidation(auditHeader) method if there is any error or
+	 * warning message then return the auditHeader. 2) based on the Record type
+	 * do following actions a) DELETE Delete the record from the main table by
+	 * using getDirectorDetailDAO().delete with parameters directorDetail,"" b)
+	 * NEW Add new record in to main table by using getDirectorDetailDAO().save
+	 * with parameters directorDetail,"" c) EDIT Update record in the main table
+	 * by using getDirectorDetailDAO().update with parameters directorDetail,""
+	 * 3) Delete the record from the workFlow table by using
+	 * getDirectorDetailDAO().delete with parameters directorDetail,"_Temp" 4)
+	 * Audit the record in to AuditHeader and AdtCustomerDirectorDetail by using
+	 * auditHeaderDAO.addAudit(auditHeader) for Work flow 5) Audit the record in
+	 * to AuditHeader and AdtCustomerDirectorDetail by using
+	 * auditHeaderDAO.addAudit(auditHeader) based on the transaction Type.
 	 * 
 	 * @param AuditHeader
 	 *            (auditHeader)
@@ -251,9 +340,11 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 			}
 		}
 
-		getDirectorDetailDAO().delete(directorDetail, "_Temp");
-		auditHeader.setAuditTranType(PennantConstants.TRAN_WF);
-		getAuditHeaderDAO().addAudit(auditHeader);
+		if (!StringUtils.equals(directorDetail.getSourceId(), PennantConstants.FINSOURCE_ID_API)) {
+			getDirectorDetailDAO().delete(directorDetail, "_Temp");
+			auditHeader.setAuditTranType(PennantConstants.TRAN_WF);
+			getAuditHeaderDAO().addAudit(auditHeader);
+		}
 
 		auditHeader.setAuditTranType(tranType);
 		auditHeader.getAuditDetail().setAuditTranType(tranType);
@@ -266,10 +357,13 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 	}
 
 	/**
-	 * doReject method do the following steps. 1) Do the Business validation by using businessValidation(auditHeader)
-	 * method if there is any error or warning message then return the auditHeader. 2) Delete the record from the
-	 * workFlow table by using getDirectorDetailDAO().delete with parameters directorDetail,"_Temp" 3) Audit the record
-	 * in to AuditHeader and AdtCustomerDirectorDetail by using auditHeaderDAO.addAudit(auditHeader) for Work flow
+	 * doReject method do the following steps. 1) Do the Business validation by
+	 * using businessValidation(auditHeader) method if there is any error or
+	 * warning message then return the auditHeader. 2) Delete the record from
+	 * the workFlow table by using getDirectorDetailDAO().delete with parameters
+	 * directorDetail,"_Temp" 3) Audit the record in to AuditHeader and
+	 * AdtCustomerDirectorDetail by using auditHeaderDAO.addAudit(auditHeader)
+	 * for Work flow
 	 * 
 	 * @param AuditHeader
 	 *            (auditHeader)
@@ -295,10 +389,12 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 	}
 
 	/**
-	 * businessValidation method do the following steps. 1) get the details from the auditHeader. 2) fetch the details
-	 * from the tables 3) Validate the Record based on the record details. 4) Validate for any business validation. 5)
-	 * for any mismatch conditions Fetch the error details from getDirectorDetailDAO().getErrorDetail with Error ID and
-	 * language as parameters. 6) if any error/Warnings then assign the to auditHeader
+	 * businessValidation method do the following steps. 1) get the details from
+	 * the auditHeader. 2) fetch the details from the tables 3) Validate the
+	 * Record based on the record details. 4) Validate for any business
+	 * validation. 5) for any mismatch conditions Fetch the error details from
+	 * getDirectorDetailDAO().getErrorDetail with Error ID and language as
+	 * parameters. 6) if any error/Warnings then assign the to auditHeader
 	 * 
 	 * @param AuditHeader
 	 *            (auditHeader)
@@ -310,6 +406,63 @@ public class DirectorDetailServiceImpl extends GenericService<DirectorDetail> im
 		auditHeader = nextProcess(auditHeader);
 		logger.debug("Leaving");
 		return auditHeader;
+	}
+
+	@Override
+	public AuditDetail doValidations(DirectorDetail directorDetail, Customer customerDetails) {
+
+		AuditDetail auditDetail = new AuditDetail();
+
+		// gender code Validation
+
+		if (genderDAO.isValidGenderCode(directorDetail.getCustGenderCode())) {
+			String[] valueParm = new String[1];
+			valueParm[0] = directorDetail.getCustGenderCode();
+			auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(new ErrorDetail("90701", valueParm)));
+		}
+
+		// salutation validation
+		int salutationByCount = salutationDAO.getSalutationByCount(directorDetail.getCustSalutationCode(),
+				directorDetail.getCustGenderCode());
+
+		if (salutationByCount <= 0) {
+			String[] valueParm = new String[2];
+			valueParm[0] = directorDetail.getCustSalutationCode();
+			valueParm[1] = directorDetail.getCustGenderCode();
+			auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(new ErrorDetail("90701", "", valueParm)));
+		}
+
+		// Id type validation
+		int docTypeCount = customerDocumentDAO.getDocTypeCount(directorDetail.getIdType());
+
+		if (docTypeCount <= 0) {
+			String[] valueParm = new String[1];
+			valueParm[0] = directorDetail.getIdType();
+			auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(new ErrorDetail("90701", "", valueParm)));
+		}
+
+		// Nationality
+		NationalityCode nationalityCodeById = nationalityCodeDAO.getNationalityCodeById(directorDetail.getNationality(),
+				"");
+		if (nationalityCodeById == null) {
+			String[] valueParm = new String[2];
+			valueParm[0] = directorDetail.getCustAddrCountry();
+			valueParm[1] = directorDetail.getCustAddrProvince();
+			auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(new ErrorDetail("90701", "", valueParm)));
+		}
+
+		// country,state,city validation
+
+		City city = cityDAO.getCityById(directorDetail.getCustAddrCountry(), directorDetail.getCustAddrProvince(),
+				directorDetail.getCustAddrCity(), "");
+		if (city == null) {
+			String[] valueParm = new String[2];
+			valueParm[0] = directorDetail.getCustAddrCountry();
+			valueParm[1] = directorDetail.getCustAddrProvince();
+			auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(new ErrorDetail("90701", "", valueParm)));
+		}
+
+		return auditDetail;
 	}
 
 }
